sequencer: fix rearranging of fixupish lines with whitespace
todo_list_rearrange_squash is skipping fixupish prefixes to find the
commit subject for rearranging the todo list of the fixupish action.

If the commit subject contains leading whitespace in the commit that
should be matched, the current implementation introduced in c44a4c650c
would not find the matching commit and ignores rearranging the
fixupish commit.  This is due to ignoring all preceding whitespace
while scanning commit subjects.

This fix ignores whitespace following fixupish prefixes while keeping
the remaining string without the fixupish prefixes intact. Hence,
whenever we found the commit subject part without fixupish prefixes,
it will retain the leading whitespace and therefore can be matched
with the commit and rearranged in the todo list.

Signed-off-by: Thomas Krennwallner <tk@postsubmeta.net>